refractor: update dashboard route to /admin
